MB-5835: Removed duplicate fields in set/delete with meta
This change also changes the rev seqno from 32 to 64 bits

Reviewed-on: http://review.couchbase.org/18063
Tested-by: Michael Wiederhold <mike@couchbase.com>

+++ couch-kvstore/couch-notifier.hh
@@ -134,7 +134,7 @@ private:
     bool configurationError;
     bool shutdown;
 
-    uint32_t seqno;
+    uint64_t seqno;
     Buffer input;
 
     /**
